# simplify test_mempool_update_performance()