### Purpose:

This benchmark fails often and takes a long time to run (for no good reason). This patch makes it run faster and (likely) more reliable in passing.

The context manager measuring time, `benchmark_runner.assert_runtime` appears to be extremely slow. The current version of this test inserts 403 blocks into the blockchain and measures each `add_block()` call individually. This takes about 3 minutes. Most of the time is spent in the context manager itself.

The current version also separates the empty first 400 blocks from the 3 blocks we put on top of it with some transactions in them, using different time limits (0.001s vs 0.15s). measuring the insertion time of the empty blocks isn't really interesting. We save a lot of time by not measuring those (thus avoiding the `benchmark_runner`'s overhead.

Also, instead of measuring the 3 block insertions independently, this patch measures them together, making the time limit 0.15 * 3 = 0.45s.

<!-- Does this PR introduce a breaking change? -->
### Current Behavior:

The test takes 3 minutes to run and is flaky.

### New Behavior:

The test takes 10 seconds to run and is more reliable.